🚚 Rename: 이동 PlaceHomeContents.tsx PlaceHome -> BDPlace Fixes: #24
